Rework values set in session on authorisation.
In order to use the auth details with other calls to google api's you
need access to the actual token, not just the hash. Instead of just
setting the details parsed out of the id_token value we set the
original auth hash merged with the hash that we parsed out of the id_token.